governance-core: change tests to embed `data.json`
See also #13
